+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 
 # Unreleased
 
+# 1.16.0 (2024-05-22)
+
 - Upgrade `@aptos-labs/aptos-cli` package to version `0.1.8`
 - [`Fix`] CLI scripts to be OS compatible with Mac, Linux and Windows
 - [`Fix`] Support generating transactions with loose types for SDK V1 backward compatibility